archive_new/requirements.txt
bleak>=0.20.0
pyqtgraph>=0.13.0
PyQt5>=5.15.0
numpy>=1.21.0
pyyaml>=6.0
qasync>=0.23.0
pyserial>=3.5
ntplib>=0.4.0
pyaudio>=0.2.11


